pyproject.toml
[project]
name = "music-bot"
version = "0.1.0"
description = "Music bot for playing music on Discord from Youtube."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.3",
    "pynacl>=1.6.0",
    "python-dotenv>=1.1.1",
    "pytubefix>=9.5.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
    "ty>=0.0.1a21",
]

[tool.ruff.lint]
select = ["ALL"]

ignore = ["D", "CPY001", "PLR6301"]

preview = true

[tool.ruff.format]

preview = true
